requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-core
httpx
pydantic
python-dotenv
pinecone
supabase
